Fix fk customer id attr in transaction.

+++ src/Sample.Enquiry.Core/Dtos/TransactionDto.cs
@@ -11,7 +11,7 @@ namespace Sample.Enquiry.Core.Dtos
         [DisplayFormat(DataFormatString = "{0:#,##0.00}")]
         public decimal Amount { get; set; }
         public string Currency { get; set; }
-        public int CustomerId { get; set; }
+        public ulong CustomerId { get; set; }
         [DisplayFormat(DataFormatString = "{0:dd/MM/yy HH:mm}")]
         public DateTime TransactionDate { get; set; }
     }
