+++ src/Sample.Enquiry.Core/Entities/Transaction.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Sample.Enquiry.Core.Events;
 using Sample.Enquiry.Core.SharedKernel;
 using Sample.Enquiry.Core.Dtos;
 
@@ -10,7 +9,7 @@ namespace Sample.Enquiry.Core.Entities
         public string Status { get; set; }
         public decimal Amount { get; set; }
         public string Currency { get; set; }
-        public int CustomerId { get; set; }
+        public ulong CustomerId { get; set; }
         public DateTime TransactionDate { get; set; }
         public TransactionDto ToDto()
         {
